+++ react/utils/gtmPush.ts
@@ -1,0 +1,5 @@
+window.dataLayer = window?.dataLayer || []
+
+export default function push(event: Record<string, unknown>) {
+  window.dataLayer.push(event)
+}