+++ CHANGELOG.md
@@ -14,6 +14,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Italian and English translations.
 
+## [3.12.0] - 2022-11-04
+
+### Added
+
+- Custom GTM dataLayer event for when user submits a review
+
+## [3.11.2] - 2022-10-20
+
+### Changed
+
+- Count stars within the `averageRatingByProductId` query
+
 ### Changed
 
 - GitHub reusable workflow and cy-runner updated to version 2